app/requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
check-manifest==0.39
defusedxml==0.6.0
devpi-client==5.0.0
devpi-common==3.4.0
diff-match-patch==20181111
Django==2.2.6
django-bootstrap4==1.0.1
django-filter==2.2.0
django-guardian==2.1.0
django-import-export==1.2.0
django-jet==1.0.8
django-object-actions==1.1.0
django-reversion==3.0.4
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==3.0.0
packaging==19.2
pkginfo==1.5.0.1
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
tox==3.14.0
urllib3==1.25.6
virtualenv==16.7.5
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
